+++ test_list2.lua
@@ -1,5 +1,5 @@
 -- Test: p_list2.lua
--- VERSION: 2.105
+-- VERSION: 2.106
 
 
 local PATH = ... and (...):match("(.-)[^%.]+$") or ""